Move css to separate folder
